+++ lib/response.js
@@ -6,7 +6,7 @@ const util = require('util')
 
 const setCookie = require('set-cookie-parser')
 
-function Response (req, onEnd, reject) {
+function Response (req, onEnd) {
   http.ServerResponse.call(this, req)
 
   this._lightMyRequest = { headers: null, trailers: {}, payloadChunks: [] }
@@ -15,24 +15,16 @@ function Response (req, onEnd, reject) {
 
   this.assignSocket(getNullSocket())
 
-  this._promiseCallback = typeof reject === 'function'
-
   let called = false
   const onEndSuccess = (payload) => {
     // no need to early-return if already called because this handler is bound `once`
     called = true
-    if (this._promiseCallback) {
-      return process.nextTick(() => onEnd(payload))
-    }
     process.nextTick(() => onEnd(null, payload))
   }
 
   const onEndFailure = (err) => {
     if (called) return
     called = true
-    if (this._promiseCallback) {
-      return process.nextTick(() => reject(err))
-    }
     process.nextTick(() => onEnd(err, null))
   }
